% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_datatable.R
\name{blast_datatable}
\alias{blast_datatable}
\title{Take a datatable and return the results of BLASTing it}
\usage{
blast_datatable(
  blast_seeds,
  save_dir,
  db_dir,
  accession_taxa_path,
  sample_size = 1000,
  wildcards = "NNNN"
)
}
\arguments{
\item{blast_seeds}{a data.frame formatted like the output from
get_blast_seeds_multi_taxa_or_db}

\item{save_dir}{a directory in which to create text files representing the
current state}

\item{db_dir}{a directory with a blast-formatted database}

\item{accession_taxa_path}{a path to an sql created by
taxonomizr::prepareDatabase()}

\item{sample_size}{the number of entries to accumulate into a fasta before
calling blastn}

\item{wildcards}{a character vector representing the number of wildcards to
discard}
}
\value{
A data.frame representing the output of blastn
}
\description{
Given a datatable with the column names of the datatable returned by
\code{\link[=get_blast_seeds]{get_blast_seeds()}}, use blastdbcmd to convert entries into
fasta files, then uses blastn to query ncbi databases for those
sequences. It compiles the results of blastn into a data.frame that it
returns. Additionally, it saves its state as text files in a specified
directory with each iteration.
}
\details{
blast_datatable uses blastdbcmd and blastn to find sequences. It samples rows
from \code{blast_seeds} and uses blastdbcmd to find a seqence that corresponds to
the accession number and forward and reverse stops recorded in the table.
blastdbcmd outputs sequences as .fasta-formatted strings, which
blast_datatable concatenates into a multi-line fasta, then passes to blastn
as an argument. blast-datatable repeats this process until no rows remain,
aggregating the results in a single data.frame.
}
\section{Saving data}{
blast_datatable writes intermediate results and metadata about the search to
local files as it goes. This mitigates the consequences of encountering an
error or experiencing other interruptions.
}

