% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blast_seeds.R
\name{get_blast_seeds}
\alias{get_blast_seeds}
\title{Function to query NCBI's primer_blast tool, append the taxonomy, and
output the results as .csvs}
\usage{
get_blast_seeds(
  forward_primer,
  reverse_primer,
  file_out_dir,
  Metabarcode_name,
  accessionTaxa,
  organism,
  mismatch = 3,
  minimum_length = 5,
  maximum_length = 500,
  primer_specificity_database = "nt",
  ...,
  return_table = TRUE
)
}
\arguments{
\item{forward_primer}{passed to primer_search, which turns it into a list of
each primer it could be based on its degenerate primers, then passes
each one in turn to NCBI}

\item{reverse_primer}{passed to primer_search, which turns it into a list of
each primer it could be based on its degenerate primers, then passes
each one in turn to NCBI}

\item{file_out_dir}{the parent directory to place the data in.
get_blast_seeds does not attempt to create a new directory, so
file_out_dir should already exist}

\item{Metabarcode_name}{used to name the subdirectory and the files. If a
directory named Metabarcode_name does not exist in file_out_dir, a
new directory will be created. get_blast_seeds appends
Metabarcode_name to the beginning of each of the four files it
generates.}

\item{accessionTaxa}{the path to sql created by taxonomizr}

\item{organism}{a vector of character vectors. Each character vector is
passed in turn to primer_search, which passes them to NCBI.
get_blast_seeds aggregates all of the results into a single file.}

\item{minimum_length}{parse_primer_hits returns a table with a product_length
column. get_blast_seeds removes each row that has a value less than
minimum_length in the product_length column.}

\item{maximum_length}{parse_primer_hits returns a table with a
product_length column. get_blast_seeds removes each row that has a
value greater than maximum_length in the product_length column}

\item{primer_specificity_database}{passed to primer_search, which passes it
to NCBI}

\item{...}{additional arguments passed to primer_search, which passes it to
NCBI}

\item{num_permutations}{passed to primer_search}

\item{hitsize}{passed to primer_search, which passes it to NCBI}

\item{evalue}{passed to primer_search, which passes it to NCBI}

\item{word_size}{passed to primer_search, which passes it to NCBI}

\item{MAX_TARGET_PER_TEMPLATE}{passed to primer_search, which passes it to
NCBI}

\item{NUM_TARGETS_WITH_PRIMERS}{passed to primer_search, which passes it to
NCBI}
}
\value{
a tibble containing the same information as the csv it generates
}
\description{
Function to query NCBI's primer_blast tool, append the taxonomy, and
output the results as .csvs
}
