% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hist_save_pdf.R
\name{make_hist_save_pdf}
\alias{make_hist_save_pdf}
\title{Function to make and save histograms as pdfs}
\usage{
make_hist_save_pdf(infile, description, file_out_dir, Metabarcode_name)
}
\description{
Function to make and save histograms as pdfs
}
