% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_state.R
\name{save_state}
\alias{save_state}
\title{Helper function to save the state of a running blast search}
\usage{
save_state(dir, output_table, unsampled_indices, too_many_ns, not_in_db)
}
\arguments{
\item{dir}{the directory to save in}

\item{output_table}{the outputs generated so far}

\item{unsampled_indices}{a vector of indices not yet sampled}

\item{too_many_ns}{a vector of indices that result
in a fasta with too many Ns}

\item{not_in_db}{a vector of indices not found in the local
database}
}
\description{
Helper function to save the state of a running blast search
}
