% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_parse_hits.R
\name{try_parse_hits}
\alias{try_parse_hits}
\title{Function to attempt parse_primer_hits and fail less loudly}
\usage{
try_parse_hits(response)
}
\arguments{
\item{response}{A response object to parse}
}
\value{
If parse_primer_hits() doesn't throw an error, the results of
parse_primer_hits(). If it does throw an error, FALSE
}
\description{
Equivalent to parse_primer_hits() if the argument is a legal
argument. If the argument is illegal, returns FALSE, which
the program can use to respond appropriately.
}
