% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrux_blast.R
\name{rcrux_blast}
\alias{rcrux_blast}
\title{wrapper function for blast_datatable that reads a data.frame from a path}
\usage{
rcrux_blast(
  seeds_path,
  db_dir,
  accession_taxa_path,
  working_dir,
  expand_vectors = TRUE,
  ...
)
}
\arguments{
\item{seeds_path}{a path to a csv from get_blast_seeds}

\item{db_dir}{a directory containing a blast-formatted database. This should
not be the path to the database as it appears in the blastn arguments
(ie it should not have the name of the database trailing).}

\item{accession_taxa_path}{a path to the accessionTaxa sql created by
taxonomizr}

\item{working_dir}{a directory in which to save partial and complete output}

\item{expand_vectors}{logical, determines whether to expand too_many_Ns
and not_in db into real tables}
}
\description{
rcrux_ blast uses \code{\link[=blast_datatable]{blast_datatable()}} to search against a blast
formatted database. It creates a directory at \code{working_dir} if one does not
already exist and generates \code{rcrux_blast_output} and \code{.rcrux_blast_save}
inside that directory. It passes \verb{<working_dir>/.rcrux_blast_save} to
\code{\link[=blast_datatable]{blast_datatable()}} as the save directory and generates files in
\code{rcrux_blast_output} recording the results of the blast.
}
